+++ editor/core/Admin.php
@@ -561,8 +561,6 @@ class Admin extends Origin
 
 	public function printNavigation(?string $missing): void
 	{
-		parent::printNavigation($missing);
-
 		if ($missing == "auth") {
 			$first = true;
 			foreach ((array)$_SESSION["pwds"] as $vendor => $servers) {